展现频次属性修改

+++ perfect-commons/src/main/java/com/perfect/dto/account/NmsAccountReportDTO.java
@@ -6,7 +6,7 @@ import java.util.Date;
 /**
  * Created by subdong on 15-7-21.
  */
-public class NmsAccountReportDTO extends AccountIdDTO{
+public class NmsAccountReportDTO extends AccountIdDTO {
     private String id;
 
     private Date date;
@@ -29,7 +29,7 @@ public class NmsAccountReportDTO extends AccountIdDTO{
 
     private Integer clickuv;    // 点击独立访客
 
-    private Integer srsur;  // 展现频次
+    private Double srsur;  // 展现频次
 
     private Double cusur;  // 独立访客点击率
 
@@ -135,11 +135,11 @@ public class NmsAccountReportDTO extends AccountIdDTO{
         this.clickuv = clickuv;
     }
 
-    public Integer getSrsur() {
+    public Double getSrsur() {
         return srsur;
     }
 
-    public void setSrsur(Integer srsur) {
+    public void setSrsur(Double srsur) {
         this.srsur = srsur;
     }
