+++ perfect-commons/src/main/java/com/perfect/dto/account/NmsGroupReportDTO.java
@@ -7,7 +7,7 @@ import static com.perfect.commons.constants.MongoEntityConstants.CAMPAIGN_ID;
 /**
  * Created by subdong on 15-7-21.
  */
-public class NmsGroupReportDTO extends AccountIdDTO{
+public class NmsGroupReportDTO extends AccountIdDTO {
 
     private String id;
 
@@ -35,7 +35,7 @@ public class NmsGroupReportDTO extends AccountIdDTO{
 
     private Integer clickuv;    // 点击独立访客
 
-    private Integer srsur;  // 展现频次
+    private Double srsur;  // 展现频次
 
     private Double cusur;  // 独立访客点击率
 
@@ -157,11 +157,11 @@ public class NmsGroupReportDTO extends AccountIdDTO{
         this.clickuv = clickuv;
     }
 
-    public Integer getSrsur() {
+    public Double getSrsur() {
         return srsur;
     }
 
-    public void setSrsur(Integer srsur) {
+    public void setSrsur(Double srsur) {
         this.srsur = srsur;
     }
